requirements.txt
streamlit>=1.34
python-docx>=1.1.0
pypdf>=4.2.0
sentence-transformers>=3.0.0
scikit-learn>=1.4
numpy>=1.26
pandas>=2.2
spacy>=3.7
rapidfuzz>=3.9
nltk>=3.9
